move commanders into agent scripted
