+++ nidup/pysc2/scripted_agents.py
@@ -1,6 +1,6 @@
 
 from pysc2.agents.base_agent import BaseAgent
-from nidup.pysc2.commanders import GameCommander, ScoutingCommander
+from nidup.pysc2.agent.scripted.commander import GameCommander, ScoutingCommander
 from nidup.pysc2.wrapper.observations import Observations
 from nidup.pysc2.information import BaseLocation
 import random
